feat: separate shouldForceSecondUpdate to util
